requirements.txt
python-telegram-bot==13.15
gspread==5.7.2
oauth2client==4.1.3
python-dotenv==1.0.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0